Improve backup success rate by adding retries on single file downloads/uploads
Add wait multiplier to S3 single part downloads/uploads to maximize success rate in case of transient failures.
Add automated retries on single file downloads.

+++ medusa/purge.py
@@ -140,9 +140,10 @@ def cleanup_obsolete_files(storage, fqdn):
     for path in paths_in_storage - paths_in_manifest:
         logging.debug("  - [{}] exists in storage, but not in manifest".format(path))
         obj = storage.storage_driver.get_blob(path)
-        nb_objects_purged += 1
-        total_purged_size += int(obj.size)
-        storage.storage_driver.delete_object(obj)
+        if obj is not None:
+            nb_objects_purged += 1
+            total_purged_size += int(obj.size)
+            storage.storage_driver.delete_object(obj)
 
     return nb_objects_purged, total_purged_size
